removed transaction when adding family columns'
closes #2293
In some cases, statements are in progress when we attempt to create the
transaction. When this occurs, the statement creation throws a fatal exception.
(fatal in the sense we decided to throw it again after catching)
Some work has to be done in the entire application in order to work with
transactions where needed.